+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/ServerProviders/AgentServerProvider.cs
@@ -1,7 +1,11 @@
 ﻿using Docker.DotNet.Models;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.Extensions.Hosting;
 using MsgPack.Serialization;
 using Newtonsoft.Json.Linq;
+using Stormancer.Server.Components;
 using Stormancer.Server.Plugins.Configuration;
+using Stormancer.Server.Plugins.GameSession.ServerPool;
 using Stormancer.Server.Plugins.Users;
 using Stormancer.Server.Secrets;
 using System;
@@ -25,6 +29,18 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
         /// List of paths in the cluster secret stores containing valid public keys for authenticating game server agents.
         /// </summary>
         public IEnumerable<string> AuthCertPaths { get; set; } = Enumerable.Empty<string>();
+    }
+
+    public class AgentPoolConfigurationSection : PoolConfiguration
+    {
+        public string Image { get; set; }
+        public Dictionary<string,string> EnvironmentVariables { get; set; }
+
+        public uint TimeoutSeconds { get; set; }
+
+        public override string type => "fromProvider";
+
+        public string provider { get; set; } = "agent";
     }
 
     internal class GameServerAgentConfiguration : IConfigurationChangedEventHandler
@@ -169,6 +185,11 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
         public bool Faulted => Fault != null;
 
         public AgentDescription Description { get; }
+
+        public float TotalCpu { get; set; }
+        public float UsedCpu { get; set; }
+        public long TotalMemory { get; set; }
+        public long UsedMemory { get; set; }
     }
     public class AgentBasedGameServerProvider : IGameServerProvider
     {
@@ -180,6 +201,7 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
             lock (_syncRoot)
             {
                 _agents.Add(agentSession.User.Id, new DockerAgent(peer, agentSession));
+                
             }
         }
 
@@ -202,6 +224,34 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
             }
         }
 
+        private async Task SubscribeContainerStatusUpdate(DockerAgent agent, CancellationToken cancellationToken)
+        {
+            await foreach(var update in GetContainerStatusUpdates(agent.Id, cancellationToken))
+            {
+                agent.TotalCpu = update.TotalCpu;
+                agent.UsedCpu = update.ReservedCpu;
+                agent.TotalMemory = update.TotalMemory;
+                agent.UsedMemory = update.ReservedMemory;
+            }
+        }
+
+        private IAsyncEnumerable<ContainerStatusUpdate> GetContainerStatusUpdates(string agentId, CancellationToken cancellationToken)
+        {
+            DockerAgent? agent;
+            lock (_syncRoot)
+            {
+                if (!_agents.TryGetValue(agentId, out agent))
+                {
+                    throw new InvalidOperationException("Agent not found");
+                }
+            }
+
+            var observable = agent.Peer.Rpc<bool, ContainerStatusUpdate>("agent.getDockerEvents", true, cancellationToken);
+
+
+            return observable.ToAsyncEnumerable();
+        }
+      
         public async IAsyncEnumerable<ContainerDescription> GetRunningContainers()
         {
             List<Task<IEnumerable<ContainerDescription>>> tasks = new List<Task<IEnumerable<ContainerDescription>>>();
@@ -240,18 +290,26 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
             return peer.RpcTask<bool, IEnumerable<ContainerDescription>>("agent.getRunningContainers", true);
         }
 
-        public Task<ContainerStartResponse> StartContainer(string agentId, string image, string containerId, float cpuQuota, int memoryQuota)
+        public Task<ContainerStartResponse> StartContainerAsync(string agentId, string image, string containerId, float cpuQuota, int memoryQuota, Dictionary<string,string> environmentVariables)
         {
             DockerAgent? agent;
-            lock(_syncRoot)
+            lock (_syncRoot)
             {
-                if(!_agents.TryGetValue(agentId,out agent))
+                if (!_agents.TryGetValue(agentId, out agent))
                 {
                     throw new InvalidOperationException("Agent not found");
                 }
             }
 
-            return agent.Peer.RpcTask<ContainerStartParameters,ContainerStartResponse>("agent.tryStartContainer", new ContainerStartParameters { containerId = containerId, cpuQuota = cpuQuota, Image = image, MemoryQuota = memoryQuota  });
+            return agent.Peer.RpcTask<ContainerStartParameters, ContainerStartResponse>("agent.tryStartContainer", new ContainerStartParameters
+            {
+                containerId = containerId,
+                cpuQuota = cpuQuota,
+                Image = image,
+                MemoryQuota = memoryQuota,
+                EnvironmentVariables = environmentVariables,
+                
+            });
         }
 
         public Task<ContainerStopResponse> StopContainer(string agentId, string containerId)
@@ -265,10 +323,10 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
                 }
             }
 
-            return agent.Peer.RpcTask<ContainerStopParameters, ContainerStopResponse>("agent.stopContainer", new ContainerStopParameters {ContainerId = containerId });
+            return agent.Peer.RpcTask<ContainerStopParameters, ContainerStopResponse>("agent.stopContainer", new ContainerStopParameters { ContainerId = containerId });
         }
 
-        public Task<GetContainerLogsResponse> GetLogs(string agentId,string containerId )
+        public IAsyncEnumerable<IEnumerable<string>> GetLogsAsync(string agentId, string containerId,bool follow, DateTime? since = null, DateTime? until = null,uint size = 0, CancellationToken cancellationToken)
         {
             DockerAgent? agent;
             lock (_syncRoot)
@@ -279,15 +337,47 @@ namespace Stormancer.Server.Plugins.GameSession.ServerProviders
                 }
             }
 
-            return agent.Peer.RpcTask<GetContainerLogsParameters, GetContainerLogsResponse>("agent.getLogs", new GetContainerLogsParameters { ContainerId = containerId });
+            var observable = agent.Peer.Rpc<GetContainerLogsParameters, IEnumerable<string>>("agent.getLogs", new GetContainerLogsParameters { 
+                ContainerId = containerId,
+                Follow = follow,
+                Since = since,
+                Until = until,
+                Size = size
+            
+            },cancellationToken);
+
+
+            return observable.ToAsyncEnumerable();
+          
         }
 
 
         public string Type => GameServerAgentConstants.TYPE;
 
-        public Task<GameServerInstance> StartServer(string id, JObject config, CancellationToken ct)
+        public Task<StartGameServerResult> TryStartServer(string id, JObject config, CancellationToken ct)
         {
-            throw new NotImplementedException();
+            var agentConfig = config.ToObject<AgentPoolConfigurationSection>();
+            var applicationInfo = await _environment.GetApplicationInfos();
+            var fed = await _environment.GetFederation();
+            var node = await _environment.GetNodeInfos();
+
+            var arguments = string.Join(" ", config.arguments ?? Enumerable.Empty<string>());
+
+            var udpTransports = node.Transports.First(t => t.Item1 == "raknet");
+            var environmentVariables = new Dictionary<string, string>
+            {
+                //startInfo.EnvironmentVariables.Add("connectionToken", token);
+                    { "Stormancer_Server_Port", server.ServerPort.ToString() },
+                    { "Stormancer_Server_ClusterEndpoints", endpoints },
+                    { "Stormancer_Server_PublishedAddresses", server.PublicIp },
+                    { "Stormancer_Server_PublishedPort", server.ServerPort.ToString() },
+                    { "Stormancer_Server_AuthenticationToken", authenticationToken },
+                    { "Stormancer_Server_Account", applicationInfo.AccountId },
+                    { "Stormancer_Server_Application", applicationInfo.ApplicationName },
+                    { "Stormancer_Server_Arguments", arguments },
+                    { "Stormancer_Server_TransportEndpoint", TransformEndpoint(udpTransports.Item2.First().Replace(":","|")) }
+                };
+    
         }
 
         public Task StopServer(string id)