+++ src/Stormancer.Plugins/Galaxy/Stormancer.Server.Plugins.Galaxy/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Updated MemoryCache implementation.
+
 3.1.1.2
 -------
 Changed