+++ src/Stormancer.Plugins/Limits/Stormancer.Server.Plugins.Limits/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Update MemoryCache implementation
+
 0.2.0.1
 ----------
 Changed